:sparkles: Parse desktop source files without the API.
Electron now loads the renderer from local build assets, owns source-file graph parsing through IPC, and keeps browser imports on the API path.
