+++ api/src/services/graphs/create-service.d.cts
@@ -1,0 +1,3 @@
+import type { SourceGraph } from "./create-service.ts"
+
+export function createSourceGraph(path: string, sourceName?: string): SourceGraph
