+++ web/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link
       rel="icon"
-      href="/favicon.ico"
+      href="./favicon.ico"
     />
     <meta
       name="viewport"
